+++ Repository/GiftVoucherAvailableRepository.php
@@ -14,8 +14,23 @@ use c975L\GiftVoucherBundle\Entity\GiftVoucherPurchased;
 
 class GiftVoucherAvailableRepository extends EntityRepository
 {
-    //Finds all GiftVoucher available
-    public function findAvailable()
+    //Finds all GiftVouchers available ordered by alphabetical order
+    public function findAllAlphabeticalOrder($number = null, $order = 'object')
+    {
+        $qb = $this->createQueryBuilder('v');
+        $qb->select('v')
+            ->where('v.suppressed is NULL')
+            ->orderBy('v.' . strtolower($order), 'ASC');
+            ;
+        if ($number !== null) {
+            $qb->setMaxResults($number);
+        }
+
+        return $qb->getQuery()->getResult();
+    }
+
+    //Finds all GiftVouchers available
+    public function findAllAvailable()
     {
         $qb = $this->createQueryBuilder('v');
         $qb->select('v')